pyproject.toml
[tool.poetry]
name = "noticias_sin_filtro_server"
version = "0.1.0"
description = "Server Backend for the 'noticias sin filtro' mobile app"
authors = ["Maria V <65676201+marylicious@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.2"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
django-environ = "^0.8.1"
Scrapy = "^2.5.1"
scrapyd = "^1.3.0"
python-scrapyd-api = "^2.1.2"
black = "^22.1.0"
pytz = "^2021.3"
Pillow = "^9.0.1"
beautifulsoup4 = "^4.10.0"
celery = "^5.2.3"
redis = "^4.1.4"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-filter = "^21.1"
parsedatetime = "^2.6"
django-colorfield = "^0.6.3"
djangorestframework-dataclasses = "^1.1.1"
django-model-utils = "^4.2.0"
django-crispy-forms = "^1.14.0"
django-datatables-view = "^1.20.0"
django-bootstrap-modal-forms = "^2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
